Fixes for PR
